Add Bare GTest Capabilities
Refactor codebase to allow for unit testing via GTest. Running
./build.sh will rebuild the build system via CMake, and will then
actually build the specified targets -- in this case, both the
disassembler and its unit tests.

I don't believe that this is the best approach, since if the
CMakeLists.txt file doesn't change, you aren't required to regenerate
the build sytsem. However, this just gets everything up and running
at the start.
